Rename const/error selector files (#320)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/320

Rename `Recoil_const.js` -> `Recoil_constSelector.js` and `Recoil_error.js` -> `Recoil_errorSelector.js` to be properly named based on what they export.

Fix `useRecoilValueLoadable-test.js` unit test.

Reviewed By: aaronabramov

Differential Revision: D22027221

fbshipit-source-id: 8ea68edb882e64a655066d0f81a6cc74b69c66f0
