+++ src/hooks/__tests__/Recoil_PublicHooks-test.js
@@ -25,7 +25,7 @@ const {act} = require('ReactTestUtils');
 
 const Queue = require('../../adt/Recoil_Queue');
 const atom = require('../../recoil_values/Recoil_atom');
-const errorSelector = require('../../recoil_values/Recoil_error');
+const errorSelector = require('../../recoil_values/Recoil_errorSelector');
 const selector = require('../../recoil_values/Recoil_selector');
 const {
   ReadsAtom,
