+++ src/recoil_values/__tests__/Recoil_selector-test.js
@@ -17,8 +17,8 @@ const React = require('React');
 const {act} = require('ReactTestUtils');
 const {useRecoilState, useRecoilValue} = require('../../Recoil_index');
 const atom = require('../Recoil_atom');
-const constSelector = require('../Recoil_const');
-const errorSelector = require('../Recoil_error');
+const constSelector = require('../Recoil_constSelector');
+const errorSelector = require('../Recoil_errorSelector');
 const {
   getRecoilValueAsLoadable,
   setRecoilValue,
